requirements.txt
grpcio==1.60.1
grpcio-tools==1.60.1
cryptography==41.0.7
flask==2.3.2
flask-inputs==0.3.0
flask-cors==3.0.10
Flask-RESTful==0.3.10
requests==2.28.2